+++ tests/offline/test_magic_dict.py
@@ -3,7 +3,7 @@ from uuid import UUID
 import pytest
 
 from id_translation.offline import MagicDict
-from id_translation.transform.types import Transformer, TransformerStop
+from id_translation.transform.types import Transformer
 
 
 def test_get(monkeypatch):
@@ -72,41 +72,37 @@ def test_transformer():
     assert subject[2] == "TWO"
     assert transformer.call_counts["try_add_missing_key"] == 1
     assert subject[2] == "TWO"
-    assert transformer.call_counts["try_add_missing_key"] == 1
+    assert transformer.call_counts["try_add_missing_key"] == 1  # hit - no increment
 
     assert subject[3] == "THREE"
     assert transformer.call_counts["try_add_missing_key"] == 2
     assert subject[3] == "THREE"
-    assert transformer.call_counts["try_add_missing_key"] == 2
+    assert transformer.call_counts["try_add_missing_key"] == 2  # hit - no increment
 
-    assert subject[4] == "<Failed: id=4>"
-    assert subject[5] == "<Failed: id=5>"
-    assert subject[6] == "<Failed: id=6>"
+    assert subject[4] == "<Failed: id=4>"  # 3rd miss
+    assert subject[4] == "<Failed: id=4>"  # 4th miss
+    assert subject[5] == "<Failed: id=5>"  # 5th miss
+    assert subject[4] == "<Failed: id=4>"  # 6th miss
+    assert subject[4] == "<Failed: id=4>"  # 7th miss
+    assert subject[6] == "<Failed: id=6>"  # 8thh miss
 
-    assert transformer.call_counts == {"update_translations": 1, "try_add_missing_key": DummyTransformer.max_try}
-    with pytest.raises(TransformerStop):
-        transformer.try_add_missing_key(-1, translations=subject)
+    assert transformer.call_counts == {"update_translations": 1, "try_add_missing_key": 8}
 
 
 class DummyTransformer(Transformer[int]):
-    max_try = 3
-
     def __init__(self):
         self.call_counts = {"update_translations": 0, "try_add_missing_key": 0}
 
-    def update_ids(self, ids):
-        pass
+    def update_ids(self, _, /):
+        raise AssertionError("called by Translator only")
 
-    def update_translations(self, translations):
+    def update_translations(self, translations, /):
         assert self.call_counts["update_translations"] == 0
         self.call_counts["update_translations"] = 1
         assert translations.get(-1) is None
 
     def try_add_missing_key(self, key, /, *, translations):
         self.call_counts["try_add_missing_key"] += 1
-
-        if self.call_counts["try_add_missing_key"] >= self.max_try:
-            raise TransformerStop()
 
         if (value := {2: "TWO", 3: "THREE"}.get(key)) is not None:
             translations[key] = value