+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Methods `MagicDict.real_get` and `real_contains`.
+
+### Changed
+- Make `Transformer` method parameters `positional`-only.
+
+### Fixed
+- The docs no longer incorrectly state the `max_fails` stops working when `default_fmt_placeholders` are in use.
+- Fix `max_fails` check when transformers are in use.
+- The `BitmaskTransformer` no longer uses missing IDs in decomposed bitmask translations.
+
+### Removed
+- The `TransformerStop` class and associated functionality (it wasn't very useful).
+
 ## [0.13.0] - 2025-04-05
 
 ### Added